Patin_Clement_4_code_dashboard_062024/frontend/requirements.txt
numpy
pandas==2.2.2
Pillow==10.4.0
plotly==5.22.0
Requests==2.32.3
segmentation_models==1.0.1
streamlit==1.32.0
streamlit_image_select==0.6.0
tensorflow==2.15.0
